fix: renamed the file and removed metadata field in the providerInfo
